requirements.txt
Django==5.2.5
Pillow==11.3.0
python-decouple==3.8
pytest-django==4.11.1
pytest-cov==6.2.1
coverage==7.10.5
